Module kilt_dip_primitives::merkle_proofs::v0

Module to deal with cross-chain Merkle proof as generated by the KILT chain.

Structs

DidMerkleProof

The Merkle proof for a KILT DID.

DipCommitmentStateProof

The state proof for a DIP commitment.

DipDidProofWithVerifiedStateRoot

A DIP proof that has had the proof header and the relaychain state verified for the provided relaychain block number.

DipDidProofWithVerifiedSubjectCommitment

A DIP proof that has had the relaychain state and the DIP commitment verified for the provided relaychain block number.

DipOriginInfo

Information, available as an origin, after the whole DIP proof has been verified.

DipRevealedDetailsAndUnverifiedDidSignature

A DIP proof whose information has been verified but that contains a cross-chain TimeBoundDidSignature that still needs verification.

DipRevealedDetailsAndVerifiedDidSignatureFreshness

A DIP proof whose information has been verified and whose signature has been verified not to be expired, but that yet does not contain information as to which of the revealed keys has generated the signature.

ParachainDipDidProof

A DIP proof submitted to a parachain consumer.

ProviderHeadStateProof

The state proof for a parachain head.

RelayDipDidProof

A DIP proof submitted to a relaychain consumer.

RelayDipDidProofWithVerifiedRelayStateRoot

A DIP proof submitted to a relaychain consumer that has had the proof header verified against a given block hash.

RevealedAccountId

The details of an account after it has been successfully verified in a Merkle proof.

RevealedDidKey

The details of a DID key after it has been successfully verified in a Merkle proof.

RevealedWeb3Name

The details of a web3name after it has been successfully verified in a Merkle proof.

TimeBoundDidSignature

A DID signature anchored to a specific block height.

Enums

DidKeyRelationship

Relationship of a key to a DID Document.

Error

RevealedDidMerkleProofLeaf

All possible Merkle leaf types that can be revealed as part of a DIP identity Merkle proof.