Function kilt_dip_primitives::state_proofs::verify_storage_value_proof

pub fn verify_storage_value_proof<StorageKey, MerkleHasher, Success>(
    storage_key: &StorageKey,
    state_root: OutputOf<MerkleHasher>,
    state_proof: impl IntoIterator<Item = Vec<u8>>
) -> Result<Success, MerkleProofError>where
    StorageKey: AsRef<[u8]>,
    MerkleHasher: Hash,
    OutputOf<MerkleHasher>: Ord,
    Success: Decode,

Verify a Merkle-based storage proof for a given storage key according to the provided state root. The generic types indicate the following:

StorageKey: defines the type of the storage key included in the proof.

MerkleHasher: defines the hashing algorithm used to calculate the Merkle root.

Success: defines the type expected to be revealed in a valid Merkle proof.