Module kilt_dip_primitives::verifier

Verifier module containing types that implement the verifier component to be deployed both on a sibling parachain and on a parent relaychain.

Re-exports

pub use parachain::DipParachainStateProofVerifierError;

pub use parachain::KiltVersionedParachainVerifier;

pub use parachain::VersionedDipParachainStateProof;

pub use relaychain::DipRelaychainStateProofVerifierError;

pub use relaychain::KiltVersionedRelaychainVerifier;

pub use relaychain::VersionedRelaychainStateProof;

Modules

parachain

Verification logic to integrate a sibling chain as a DIP provider.

relaychain

Verification logic to integrate a child chain as a DIP provider.